# Improve fp64ify perf