--- a/src/core/index.js
+++ b/src/core/index.js
@@ -92,7 +92,7 @@ import {clamp} from './utils/scale-utils';
 import {flatten, countVertices, flattenVertices, fillArray} from './utils/flatten';
 // TODO - just expose as layer methods instead?
 import {enable64bitSupport} from './utils/fp64';
-import {fp64ify} from './utils/fp64';
+import {fp64ify, fp64LowPart} from './utils/fp64';
 
 export const experimental = {
   ViewportControls: Controller,
@@ -145,5 +145,6 @@ export const experimental = {
   fillArray,
 
   enable64bitSupport,
-  fp64ify
+  fp64ify,
+  fp64LowPart
 };