Avoid creating intermediate arrays.

Node:

├─ fp64#fp64ify(x)[1]: 21.3M iterations/s (0.03s elapsed)
├─ fp64#fp64LowPart(x): 39.5M iterations/s (0.03s elapsed)

Browser:

├─ fp64#fp64ify(x)[1]: 30.6M iterations/s (0.04s elapsed)
├─ fp64#fp64LowPart(x): 43.9M iterations/s (0.03s elapsed)

Verified: layer-browser